+++ src/components/ProposalDetailView.tsx
