+++ src/components/ProposalItem.tsx
