+++ src/components/Treasure.tsx
