+++ src/utils/formatBalance.ts
